Reorganizacion de carpetas de handlers, helpers y routes del back
