+++ api/src/utils/Videogames/getInfoGames.js
@@ -1,7 +1,7 @@
 require('dotenv').config();
 const axios = require('axios');
 const {API_KEY} = process.env;
-const {Videogame, Genre} = require('../db');
+const {Videogame, Genre} = require('../../db');
 const URL = "https://api.rawg.io/api/games";
