+++ api/src/utils/genres/GetInfoGenres.js
